Commited helper.go demonstrating packages in Go!
