+++ booking-app/main.go
@@ -82,15 +82,6 @@ func getFirstNames() []string {
 	return firstNames
 }
 
-// Validates user input
-func validateUserInput(firstName string, lastName string, email string, userTickets uint) (bool, bool, bool) {
-	isValidName := len(firstName) >= 2 && len(lastName) >= 2
-	isValidEmail := strings.Contains(email, "@")
-	isValidTicketNumber := userTickets > 0 && userTickets <= remainingTickets
-
-	return isValidName, isValidEmail, isValidTicketNumber
-}
-
 // Get user input
 func getUserInput() (string, string, string, uint) {
 	var firstName string
